+++ src/jsontree.js
@@ -4,7 +4,7 @@
  * A lightweight JavaScript library that generates customizable tree views to better visualize JSON data.
  * 
  * @file        jsontree.js
- * @version     v0.3.0
+ * @version     v0.4.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2024
@@ -15,6 +15,7 @@
     var // Variables: Constructor Parameters
         _parameter_Document = null,
         _parameter_Window = null,
+        _parameter_Navigator = null,
         _parameter_Math = null,
         _parameter_JSON = null,
 
@@ -140,28 +141,42 @@
      */
 
     function renderControlTitleBar( bindingOptions ) {
-        if ( bindingOptions.showTitle || bindingOptions.showTitleTreeControls ) {
-            var titleBar = createElement( bindingOptions.currentView.element, "div", "title-bar" );
+        if ( bindingOptions.showTitle || bindingOptions.showTitleTreeControls || bindingOptions.showTitleCopyButton ) {
+            var titleBar = createElement( bindingOptions.currentView.element, "div", "title-bar" ),
+                controls = createElement( titleBar, "div", "controls" );
         
             if ( bindingOptions.showTitle ) {
-                createElementWithHTML( titleBar, "div", "title", bindingOptions.titleText );
+                createElementWithHTML( titleBar, "div", "title", bindingOptions.titleText, controls );
+            }
+
+            if ( bindingOptions.showTitleCopyButton ) {
+                var copy = createElementWithHTML( controls, "button", "copy-all", _configuration.copyAllButtonText );
+
+                copy.onclick = function() {
+                    var copyData = _parameter_JSON.stringify( _elements_Data[ bindingOptions.currentView.element.id ].data );
+
+                    _parameter_Navigator.clipboard.writeText( copyData );
+
+                    fireCustomTrigger( bindingOptions.onCopyAll, copyData );
+                };
             }
 
             if ( bindingOptions.showTitleTreeControls ) {
-                var controls = createElement( titleBar, "div", "controls" ),
-                    openAll = createElementWithHTML( controls, "button", "openAll", _configuration.openAllButtonText ),
+                var openAll = createElementWithHTML( controls, "button", "openAll", _configuration.openAllButtonText ),
                     closeAll = createElementWithHTML( controls, "button", "closeAll", _configuration.closeAllButtonText );
 
                 openAll.onclick = function() {
                     bindingOptions.showAllAsClosed = false;
 
                     renderControlContainer( bindingOptions );
+                    fireCustomTrigger( bindingOptions.onOpenAll, bindingOptions.currentView.element );
                 };
 
                 closeAll.onclick = function() {
                     bindingOptions.showAllAsClosed = true;
 
                     renderControlContainer( bindingOptions );
+                    fireCustomTrigger( bindingOptions.onCloseAll, bindingOptions.currentView.element );
                 };
             }
         }
@@ -180,10 +195,10 @@
             arrow = bindingOptions.showArrowToggles ? createElement( objectTypeTitle, "div", "down-arrow" ) : null,
             propertyCount = renderObjectValues( arrow, objectTypeContents, bindingOptions, data );
 
-        createElementWithHTML( objectTypeTitle, "span", "object", _configuration.objectText );
+        createElementWithHTML( objectTypeTitle, "span", bindingOptions.showValueColors ? "object" : _string.empty, _configuration.objectText );
 
         if ( bindingOptions.showCounts && propertyCount > 0 ) {
-            createElementWithHTML( objectTypeTitle, "span", "object count", "{" + propertyCount + "}" );
+            createElementWithHTML( objectTypeTitle, "span", bindingOptions.showValueColors ? "object count" : "count", "{" + propertyCount + "}" );
         }
     }
 
@@ -192,12 +207,12 @@
             objectTypeContents = createElement( container, "div", "object-type-contents" ),
             arrow = bindingOptions.showArrowToggles ? createElement( objectTypeTitle, "div", "down-arrow" ) : null;
 
-        createElementWithHTML( objectTypeTitle, "span", "array", _configuration.arrayText );
+        createElementWithHTML( objectTypeTitle, "span", bindingOptions.showValueColors ? "array" : _string.empty, _configuration.arrayText );
 
         renderArrayValues( arrow, objectTypeContents, bindingOptions, data );
 
         if ( bindingOptions.showCounts ) {
-            createElementWithHTML( objectTypeTitle, "span", "array count", "[" + data.length + "]" );
+            createElementWithHTML( objectTypeTitle, "span", bindingOptions.showValueColors ? "array count" : "count", "[" + data.length + "]" );
         }
     }
 
@@ -263,7 +278,7 @@
 
         if ( !isDefined( value ) ) {
             if ( !bindingOptions.ignoreNullValues ) {
-                valueElement = createElementWithHTML( objectTypeValue, "span", "null", "null" );
+                valueElement = createElementWithHTML( objectTypeValue, "span", bindingOptions.showValueColors ? "null" : _string.empty, "null" );
 
                 createComma( bindingOptions, objectTypeValue, isLastItem );
 
@@ -273,7 +288,7 @@
 
         } else if ( isDefinedFunction( value ) ) {
             if ( !bindingOptions.ignoreFunctionValues ) {
-                valueElement = createElementWithHTML( objectTypeValue, "span", "function", getFunctionName( value ) );
+                valueElement = createElementWithHTML( objectTypeValue, "span", bindingOptions.showValueColors ? "function" : _string.empty, getFunctionName( value ) );
             
                 createComma( bindingOptions, objectTypeValue, isLastItem );
 
@@ -282,32 +297,32 @@
             }
 
         } else if ( isDefinedBoolean( value ) ) {
-            valueElement = createElementWithHTML( objectTypeValue, "span", "boolean", value );
+            valueElement = createElementWithHTML( objectTypeValue, "span", bindingOptions.showValueColors ? "boolean" : _string.empty, value );
             
             createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedDecimal( value ) ) {
-            valueElement = createElementWithHTML( objectTypeValue, "span", "decimal", value );
+            valueElement = createElementWithHTML( objectTypeValue, "span", bindingOptions.showValueColors ? "decimal" : _string.empty, value );
             
             createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedNumber( value ) ) {
-            valueElement = createElementWithHTML( objectTypeValue, "span", "number", value );
+            valueElement = createElementWithHTML( objectTypeValue, "span", bindingOptions.showValueColors ? "number" : _string.empty, value );
             
             createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedString( value ) ) {
-            valueElement = createElementWithHTML( objectTypeValue, "span", "string", bindingOptions.showStringQuotes ? "\"" + value + "\"" : value );
+            valueElement = createElementWithHTML( objectTypeValue, "span", bindingOptions.showValueColors ? "string" : _string.empty, bindingOptions.showStringQuotes ? "\"" + value + "\"" : value );
             
             createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedDate( value ) ) {
-            valueElement = createElementWithHTML( objectTypeValue, "span", "date", getCustomFormattedDateTimeText( value, bindingOptions.dateTimeFormat ) );
+            valueElement = createElementWithHTML( objectTypeValue, "span", bindingOptions.showValueColors ? "date" : _string.empty, getCustomFormattedDateTimeText( value, bindingOptions.dateTimeFormat ) );
 
             createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedObject( value ) && !isDefinedArray( value ) ) {
-            var objectTitle = createElement( objectTypeValue, "span", "object" ),
+            var objectTitle = createElement( objectTypeValue, "span", bindingOptions.showValueColors ? "object" : _string.empty ),
                 objectTypeContents = createElement( objectTypeValue, "div", "object-type-contents" ),
                 propertyCount = renderObjectValues( arrow, objectTypeContents, bindingOptions, value );
 
@@ -320,7 +335,7 @@
             createComma( bindingOptions, objectTitle, isLastItem );
 
         } else if ( isDefinedArray( value ) ) {
-            var arrayTitle = createElement( objectTypeValue, "span", "array" ),
+            var arrayTitle = createElement( objectTypeValue, "span", bindingOptions.showValueColors ? "array" : _string.empty ),
                 arrayTypeContents = createElement( objectTypeValue, "div", "object-type-contents" );
 
             createElementWithHTML( arrayTitle, "span", "title", _configuration.arrayText );
@@ -333,7 +348,7 @@
             renderArrayValues( arrow, arrayTypeContents, bindingOptions, value );
 
         } else {
-            valueElement = createElementWithHTML( objectTypeValue, "span", "unknown", value.toString() );
+            valueElement = createElementWithHTML( objectTypeValue, "span", bindingOptions.showValueColors ? "unknown" : _string.empty, value.toString() );
 
             createComma( bindingOptions, objectTypeValue, isLastItem );
         }
@@ -437,6 +452,8 @@
         options.ignoreFunctionValues = getDefaultBoolean( options.ignoreFunctionValues, false );
         options.reverseArrayValues = getDefaultBoolean( options.reverseArrayValues, false );
         options.addArrayIndexPadding = getDefaultBoolean( options.addArrayIndexPadding, false );
+        options.showTitleCopyButton = getDefaultBoolean( options.showTitleCopyButton, false );
+        options.showValueColors = getDefaultBoolean( options.showValueColors, true );
 
         options = buildAttributeOptionStrings( options );
         options = buildAttributeOptionCustomTriggers( options );
@@ -455,6 +472,9 @@
         options.onRenderComplete = getDefaultFunction( options.onRenderComplete, null );
         options.onValueClick = getDefaultFunction( options.onValueClick, null );
         options.onRefresh = getDefaultFunction( options.onRefresh, null );
+        options.onCopyAll = getDefaultFunction( options.onCopyAll, null );
+        options.onOpenAll = getDefaultFunction( options.onOpenAll, null );
+        options.onCloseAll = getDefaultFunction( options.onCloseAll, null );
 
         return options;
     }
@@ -834,6 +854,7 @@
         _configuration.arrayText = getDefaultString( _configuration.arrayText, "array" );
         _configuration.closeAllButtonText = getDefaultString( _configuration.closeAllButtonText, "Close All" );
         _configuration.openAllButtonText = getDefaultString( _configuration.openAllButtonText, "Open All" );
+        _configuration.copyAllButtonText = getDefaultString( _configuration.copyAllButtonText, "Copy All" );
     }
 
 
@@ -853,7 +874,7 @@
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "0.3.0";
+        return "0.4.0";
     };
 
 
@@ -863,9 +884,10 @@
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      */
 
-    ( function ( documentObject, windowObject, mathObject, jsonObject ) {
+    ( function ( documentObject, windowObject, navigatorObject, mathObject, jsonObject ) {
         _parameter_Document = documentObject;
         _parameter_Window = windowObject;
+        _parameter_Navigator = navigatorObject;
         _parameter_Math = mathObject;
         _parameter_JSON = jsonObject;
 
@@ -879,5 +901,5 @@
             _parameter_Window.$jsontree = this;
         }
 
-    } ) ( document, window, Math, JSON );
+    } ) ( document, window, navigator, Math, JSON );
 } )();